Update eng leagues clubs

+++ database/migrations/2024_02_20_181504_create_accounts_table.php
@@ -11,11 +11,11 @@ return new class extends Migration
         Schema::create('accounts', function (Blueprint $table) {
             $table->id();
             $table->integer('club_id');
-            $table->integer('balance');
-            $table->integer('future_balance');
-            $table->integer('allowed_debt');
-            $table->integer('transfer_budget');
-            $table->integer('salaries_yearly_budget');
+            $table->bigInteger('balance');
+            $table->bigInteger('future_balance');
+            $table->bigInteger('allowed_debt');
+            $table->bigInteger('transfer_budget');
+            $table->bigInteger('salaries_yearly_budget');
         });
     }
